+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 - Updated minimum Grafana version to 10.1.0 (#72)
 - Added get dashboards meta request (#70)
-- Added data source health check (#71, #72)
+- Added data source health check (#71, #73)
 
 ## 3.0.0 (2024-07-11)